Move make_icloudapi_tests to testtools
