+++ flocker/node/agents/functional/test_ebs.py
@@ -35,10 +35,12 @@ from .._logging import (
     CREATE_VOLUME_FAILURE, AWS_ACTION, VOLUME_BUSY_MESSAGE,
 )
 from ..test.test_blockdevice import (
-    make_iblockdeviceapi_tests, make_iprofiledblockdeviceapi_tests,
+    make_iprofiledblockdeviceapi_tests,
+)
+from ..testtools import (
+    make_iblockdeviceapi_tests,
     make_icloudapi_tests,
 )
-
 from ..test.blockdevicefactory import (
     InvalidConfig, ProviderType, get_blockdevice_config,
     get_blockdeviceapi_with_cleanup, get_device_allocation_unit,
