+++ flocker/node/agents/testtools/_blockdevice.py
@@ -16,20 +16,27 @@ from eliot import (
     write_traceback,
 )
 
+from testtools.matchers import AllMatch, IsInstance
+
+from twisted.internet import reactor
+from twisted.python.components import proxyForInterface
 from twisted.python.filepath import FilePath
 
+from zope.interface import implementer
 from zope.interface.verify import verifyObject
 
-from ....testtools import TestCase
+from ....testtools import TestCase, AsyncTestCase
 
 from ..blockdevice import (
     AlreadyAttachedVolume,
     BlockDeviceVolume,
     IBlockDeviceAPI,
+    ICloudAPI,
+    UnattachedVolume,
     UnknownVolume,
+    _SyncToThreadedAsyncCloudAPIAdapter,
     allocated_size,
     get_blockdevice_volume,
-    UnattachedVolume,
 )
 from ..loopback import check_allocatable_size
 
@@ -746,3 +753,69 @@ def mountroot_for_test(test_case):
     mountroot.makedirs()
     test_case.addCleanup(umount_all, mountroot)
     return mountroot
+
+
+def make_icloudapi_tests(
+        blockdevice_api_factory,
+):
+    """
+    :param blockdevice_api_factory: A factory which will be called
+        with the generated ``TestCase`` during the ``setUp`` for each
+        test and which should return a provider of both ``IBlockDeviceAPI``
+        and ``ICloudAPI`` to be tested.
+
+    :returns: A ``TestCase`` with tests that will be performed on the
+       supplied ``IBlockDeviceAPI``/``ICloudAPI`` provider.
+    """
+    class Tests(AsyncTestCase):
+        def setUp(self):
+            super(Tests, self).setUp()
+            self.api = blockdevice_api_factory(test_case=self)
+            self.this_node = self.api.compute_instance_id()
+            self.async_cloud_api = _SyncToThreadedAsyncCloudAPIAdapter(
+                _reactor=reactor, _sync=self.api,
+                _threadpool=reactor.getThreadPool())
+
+        def test_interface(self):
+            """
+            The result of the factory provides ``ICloudAPI``.
+            """
+            self.assertTrue(verifyObject(ICloudAPI, self.api))
+
+        def test_current_machine_is_live(self):
+            """
+            The machine running the test is reported as alive.
+            """
+            d = self.async_cloud_api.list_live_nodes()
+            d.addCallback(lambda live:
+                          self.assertIn(self.api.compute_instance_id(), live))
+            return d
+
+        def test_list_live_nodes(self):
+            """
+            ``list_live_nodes`` returns an iterable of unicode values.
+            """
+            live_nodes = self.api.list_live_nodes()
+            self.assertThat(live_nodes, AllMatch(IsInstance(unicode)))
+
+    return Tests
+
+
+@implementer(ICloudAPI)
+class FakeCloudAPI(proxyForInterface(IBlockDeviceAPI)):
+    """
+    Wrap a ``IBlockDeviceAPI`` and also provide ``ICloudAPI``.
+    """
+    def __init__(self, block_api, live_nodes=()):
+        """
+        @param block_api: ``IBlockDeviceAPI`` to wrap.
+        @param live_nodes: Live nodes beyond the current one.
+        """
+        self.original = block_api
+        self.live_nodes = live_nodes
+
+    def list_live_nodes(self):
+        return [self.compute_instance_id()] + list(self.live_nodes)
+
+    def start_node(self, node_id):
+        return
